% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{accessors}
\alias{accessors}
\alias{forecasts}
\alias{observations}
\title{Accessing original forecast and observation data for triptych objects}
\usage{
forecasts(x, ...)

observations(x, ...)
}
\arguments{
\item{x}{An object from which the relevant information should be extracted.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
For \code{forecasts()}: A tibble of the original forecasts in long format.

For \code{observations()}: A vector of the observations.
}
\description{
Accessing original forecast and observation data for triptych objects
}
\examples{
data(ex_binary, package = "triptych")
tr <- triptych(ex_binary)

forecasts(tr)
observations(tr) 

}
\seealso{
\code{\link[=estimates]{estimates()}}, \code{\link[=regions]{regions()}}
}
