% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{add_confidence}
\alias{add_confidence}
\title{Adding confidence regions}
\usage{
add_confidence(x, level, ...)
}
\arguments{
\item{x}{An object to which a confidence region should be added.}

\item{level}{A single value for the level of confidence.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Adding confidence regions
}
