% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{add_consistency}
\alias{add_consistency}
\title{Adding consistency regions}
\usage{
add_consistency(x, level, ...)
}
\arguments{
\item{x}{An object to which a consistency region should be added.}

\item{level}{A single value for the level of confidence.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Adding consistency regions
}
