% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcbdsc.R, R/murphy.R, R/observations.R,
%   R/reliability.R, R/roc.R
\name{estimates.triptych_mcbdsc}
\alias{estimates.triptych_mcbdsc}
\alias{estimates.triptych_murphy}
\alias{estimates}
\alias{estimates.triptych_reliability}
\alias{estimates.triptych_roc}
\title{Accessing diagnostic estimate data}
\usage{
\method{estimates}{triptych_mcbdsc}(x, ...)

\method{estimates}{triptych_murphy}(x, at = NULL, ...)

estimates(x, at, ...)

\method{estimates}{triptych_reliability}(x, at = NULL, ...)

\method{estimates}{triptych_roc}(x, at = NULL, p1 = mean(observations(x)), ...)
}
\arguments{
\item{x}{An object from which the estimate information should be extracted.}

\item{...}{Additional arguments passed to other methods.}

\item{at}{A vector of thresholds where \code{x} should be evaluated.}

\item{p1}{The unconditional event probability. Used in combination with \code{at}
to determine the diagonal lines along which to determine the estimate.}
}
\value{
A tibble with the relevant information describing
the diagnostic estimate
(Murphy curve, reliability curve, ROC curve, score decomposition)
for all supplied forecasting methods.
}
\description{
Accessing diagnostic estimate data
}
\examples{
data(ex_binary, package = "triptych")
tr <- triptych(ex_binary)

estimates(tr$murphy)
estimates(tr$reliability)
estimates(tr$roc)
estimates(tr$mcbdsc)

}
\seealso{
\code{\link[=regions]{regions()}}, \code{\link[=forecasts]{forecasts()}}, \code{\link[=observations]{observations()}}
}
