% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_binary}
\alias{ex_binary}
\title{Example data set of binary observations and probability forecasts}
\format{
\subsection{\code{ex_binary}}{

A data frame with 1,000 rows and 11 columns:
\describe{
\item{y}{observations}
\item{X01}{forecasts as specified above}
\item{X02}{forecasts as specified above}
\item{X03}{forecasts as specified above}
\item{X04}{forecasts as specified above}
\item{X05}{forecasts as specified above}
\item{X06}{forecasts as specified above}
\item{X07}{forecasts as specified above}
\item{X08}{forecasts as specified above}
\item{X09}{forecasts as specified above}
\item{X10}{forecasts as specified above}
}
}
}
\usage{
ex_binary
}
\description{
A more detailed explanation of the simulation setting...
}
\keyword{datasets}
