% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_binary_nested}
\alias{ex_binary_nested}
\title{Example data set with binary observations and forecasts from nested information sets}
\format{
\subsection{\code{ex_binary_nested}}{

A data frame with 1,000 rows and 4 columns:
\describe{
\item{y}{observations}
\item{X0}{forecasts from largest information set}
\item{X1}{forecasts from second largest information set}
\item{X2}{forecasts from third largest information set}
\item{X3}{forecasts from smallest information set}
}
}
}
\usage{
ex_binary_nested
}
\description{
A more detailed explanation of the simulation setting...
}
\keyword{datasets}
