% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcbdsc.R
\name{mcbdsc}
\alias{mcbdsc}
\alias{as_mcbdsc}
\title{Evaluation of forecasts using score decompositions}
\usage{
mcbdsc(x, y = NULL, score = "Brier_score", ...)

as_mcbdsc(x, r)
}
\arguments{
\item{x}{A data frame, list, matrix, or other object that can be coerced to a tibble. Contains numeric predictions.}

\item{y}{A numeric vector of observations. If missing, a column with name "y" must be present in the tibble coerced from \code{x}.}

\item{score}{A string specifying the score function.
One of: \code{"Brier_score"} (default), \code{"log_score"}, \code{"MR_score"}.}

\item{...}{Unused.}

\item{r}{A reference triptych_mcbdsc object whose attributes are used for casting.}
}
\value{
A \code{triptych_mcbdsc} object, that is a \code{vctrs_vctr} subclass, and has
a length equal to number of prediction methods supplied in \code{x}. Each entry
is named according to the corresponding prediction method,
and contains a list of named objects:
\itemize{
\item \code{estimate}: A data frame of the score decomposition.
\item \code{region}: Either an empty list, or a (yet to be determined) object added
by \code{\link[=add_confidence]{add_confidence()}}.
\item \code{x}: The numeric vector of original predictions.
Access is most convenient through \code{\link[=estimates]{estimates()}}, \code{\link[=regions]{regions()}}, and \code{\link[=forecasts]{forecasts()}}.
}
}
\description{
A score decomposition splits the mean score into the three components of
miscalibration (MCB), discrimination (DSC), and uncertainty (UNC). Plotting
the DSC component against the MCB component allows for a quick visual
inspection of predictive performance for many prediction methods.
}
\examples{
# Construction
predictions <- matrix(runif(300), ncol = 3)
colnames(predictions) <- c("Method_1", "Method_2", "Method_3")
observations <- rbinom(100, 1, predictions[, 1])
md1 <- mcbdsc(predictions, observations)

pred_obs <- tibble::tibble(M1 = runif(100), y = rbinom(100, 1, M1))
md2 <- mcbdsc(pred_obs)

# Visualization
autoplot(md1)

}
\seealso{
Accessors: \code{\link[=estimates]{estimates()}}, \code{\link[=regions]{regions()}}, \code{\link[=forecasts]{forecasts()}}

Adding uncertainty quantification: \code{\link[=add_confidence]{add_confidence()}}

Visualization: \code{\link[=plot.triptych]{plot.triptych()}}, \code{\link[=autoplot.triptych]{autoplot.triptych()}}
}
