% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/murphy.R
\name{murphy}
\alias{murphy}
\alias{as_murphy}
\title{Evaluation of forecasts using Murphy curves}
\usage{
murphy(x, y = NULL, ref = NULL, ...)

as_murphy(x, r)
}
\arguments{
\item{x}{A data frame, list, matrix, or other object that can be coerced to a tibble. Contains numeric forecasts.}

\item{y}{A numeric vector of observations. If missing, a column with name "y" must be present in the tibble coerced from \code{x}.}

\item{ref}{A numeric vector of reference forecasts. Can also be supplied as a column with name "ref" in the tibble coerced from \code{x}.}

\item{...}{Unused.}

\item{r}{A reference triptych_murphy object whose attributes are used for casting.}
}
\value{
A \code{triptych_murphy} object, that is a \code{vctrs_vctr} subclass, and has
a length equal to number of forecasting methods supplied in \code{x}. Each entry
is named according to the corresponding forecasting method,
and contains a list of named objects:
* \code{estimate}: A data frame with the threshold and corresponding mean score values.
* \code{region}: Either an empty list, or a data frame of point confidence intervals
added by \code{\link[=add_confidence]{add_confidence()}}.
* \code{x}: The numeric vector of original forecasts.
Access is most convenient through \code{\link[=estimates]{estimates()}}, \code{\link[=regions]{regions()}}, and \code{\link[=forecasts]{forecasts()}}.
}
\description{
A Murphy curve visualizes economic utility by displaying the mean elementary
scores across all threshold values.
}
\examples{
# Construction
forecasts <- matrix(runif(300), ncol = 3)
colnames(forecasts) <- c("Method_1", "Method_2", "Method_3")
observations <- rbinom(100, 1, forecasts[, 1])
mur1 <- murphy(forecasts, observations)

pred_obs <- tibble::tibble(M1 = runif(100), y = rbinom(100, 1, M1))
mur2 <- murphy(pred_obs)

# Visualization
autoplot(mur1)

library(ggplot2)
df_est <- estimates(mur1)
ggplot(df_est) +
  geom_path(aes(x = knot, y = mean_score, col = forecast))

}
\seealso{
Accessors: \code{\link[=estimates]{estimates()}}, \code{\link[=regions]{regions()}}, \code{\link[=forecasts]{forecasts()}}

Adding uncertainty quantification: \code{\link[=add_confidence]{add_confidence()}}

Visualization: \code{\link[=plot.triptych]{plot.triptych()}}, \code{\link[=autoplot.triptych]{autoplot.triptych()}}
}
