% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/murphy.R, R/reliability.R, R/resampling.R,
%   R/roc.R
\name{resampling_cases.triptych_murphy}
\alias{resampling_cases.triptych_murphy}
\alias{resampling_cases.triptych_reliability}
\alias{resampling_cases}
\alias{resampling_cases.triptych_roc}
\title{Bootstrap case resampling for triptych objects}
\usage{
\method{resampling_cases}{triptych_murphy}(x, level = 0.9, n_boot = 1000, ...)

\method{resampling_cases}{triptych_reliability}(x, level = 0.9, n_boot = 1000, ...)

resampling_cases(x, level = 0.9, n_boot = 1000, ...)

\method{resampling_cases}{triptych_roc}(x, level = 0.9, n_boot = 1000, ...)
}
\arguments{
\item{x}{One of the triptych objects.}

\item{level}{A single value that determines which quantiles of
the bootstrap sample to return. These quantiles envelop \code{level * n_boot}
bootstrap draws.}

\item{n_boot}{The number of bootstrap samples.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
A list of tibbles that contain the information to draw confidence regions.
The length is equal to the number of forecasting methods in \code{x}.
}
\description{
Bootstrap case resampling for triptych objects
}
\examples{
# Triptych construction
forecasts <- matrix(runif(300), ncol = 3)
colnames(forecasts) <- c("Method_1", "Method_2", "Method_3")
observations <- rbinom(100, 1, forecasts[, 1])
tr <- triptych(forecasts, observations)

resampling_cases(tr$murphy, n_boot = 50)
resampling_cases(tr$reliability, n_boot = 50)
resampling_cases(tr$roc, n_boot = 50)

}
