% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{roc}
\alias{roc}
\alias{as_roc}
\title{Evaluation of forecasts using ROC curves}
\usage{
roc(x, y = NULL, convex = TRUE, ...)

as_roc(x, r)
}
\arguments{
\item{x}{A data frame, list, matrix, or other object that can be coerced to a tibble. Contains numeric predictions.}

\item{y}{A numeric vector of observations. If missing, a column with name "y" must be present in the tibble coerced from \code{x}.}

\item{convex}{A boolean value indicating whether to calculate the convex hull
or the raw ROC diagnostic.}

\item{...}{Unused.}

\item{r}{A reference triptych_mcbdsc object whose attributes are used for casting.}
}
\value{
A \code{triptych_roc} object, that is a \code{vctrs_vctr} subclass, and has
a length equal to number of prediction methods supplied in \code{x}. Each entry
is named according to the corresponding prediction method,
and contains a list of named objects:
\itemize{
\item \code{estimate}: A data frame of hit rates and false rates.
\item \code{region}: Either an empty list, or a data frame of pointwise
confidence intervals (along diagonal lines with slope \eqn{-\pi_0/\pi_1})
added by \code{\link[=add_confidence]{add_confidence()}}.
\item \code{x}: The numeric vector of original predictions.
Access is most convenient through \code{\link[=estimates]{estimates()}}, \code{\link[=regions]{regions()}}, and \code{\link[=forecasts]{forecasts()}}.
}
}
\description{
A ROC curve visualizes discrimination ability by displaying the hit rate against
the false alarm rate for all threshold values.
}
\examples{
# Construction
predictions <- matrix(runif(300), ncol = 3)
colnames(predictions) <- c("Method_1", "Method_2", "Method_3")
observations <- rbinom(100, 1, predictions[, 1])
roc1 <- roc(predictions, observations)

pred_obs <- tibble::tibble(M1 = runif(100), y = rbinom(100, 1, M1))
roc2 <- roc(pred_obs)

# Visualization
autoplot(roc1)

library(ggplot2)
df_est <- estimates(roc1)
ggplot(df_est) +
  geom_segment(aes(x = 0, y = 0, xend = 1, yend = 1)) +
  geom_path(aes(x = FAR, y = HR, col = forecast)) +
  labs(x = "False alarm rate", y = "Hit rate")

}
\seealso{
Accessors: \code{\link[=estimates]{estimates()}}, \code{\link[=regions]{regions()}}, \code{\link[=forecasts]{forecasts()}}

Adding uncertainty quantification: \code{\link[=add_confidence]{add_confidence()}}

Visualization: \code{\link[=plot.triptych]{plot.triptych()}}, \code{\link[=autoplot.triptych]{autoplot.triptych()}}
}
