% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcbdsc.R, R/murphy.R, R/observations.R,
%   R/reliability.R, R/roc.R
\name{regions.triptych_mcbdsc}
\alias{regions.triptych_mcbdsc}
\alias{regions.triptych_murphy}
\alias{regions}
\alias{regions.triptych_reliability}
\alias{regions.triptych_roc}
\title{Accessing confidence/consistency region data}
\usage{
\method{regions}{triptych_mcbdsc}(x, ...)

\method{regions}{triptych_murphy}(x, ...)

regions(x, ...)

\method{regions}{triptych_reliability}(x, ...)

\method{regions}{triptych_roc}(x, ...)
}
\arguments{
\item{x}{An object from which the region information should be extracted.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
A tibble with the relevant information for the
uncertainty quantification of the chosen diagnostic
(Murphy curve, reliability curve, ROC curve, score decomposition)
for all supplied forecasting methods.
}
\description{
Accessing confidence/consistency region data
}
\seealso{
\code{\link[=estimates]{estimates()}}, \code{\link[=forecasts]{forecasts()}}, \code{\link[=observations]{observations()}}
}
