% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/murphy.R, R/reliability.R, R/resampling.R,
%   R/roc.R
\name{resampling_Bernoulli.triptych_murphy}
\alias{resampling_Bernoulli.triptych_murphy}
\alias{resampling_Bernoulli.triptych_reliability}
\alias{resampling_Bernoulli}
\alias{resampling_Bernoulli.triptych_roc}
\title{Bootstrap observation resampling for triptych objects}
\usage{
\method{resampling_Bernoulli}{triptych_murphy}(x, level = 0.9, n_boot = 1000, ...)

\method{resampling_Bernoulli}{triptych_reliability}(
  x,
  level = 0.9,
  n_boot = 1000,
  position = c("diagonal", "estimate"),
  ...
)

resampling_Bernoulli(x, level = 0.9, n_boot = 1000, ...)

\method{resampling_Bernoulli}{triptych_roc}(x, level = 0.9, n_boot = 1000, ...)
}
\arguments{
\item{x}{One of the triptych objects.}

\item{level}{A single value that determines which quantiles of
the bootstrap sample to return. These quantiles envelop \code{level * n_boot}
bootstrap draws.}

\item{n_boot}{The number of bootstrap samples.}

\item{...}{Additional arguments passed to other methods.}

\item{position}{Either \code{"estimate"} for confidence regions, or \code{"diagonal"}
for consistency regions.}
}
\value{
A list of tibbles that contain the information to draw confidence regions.
The length is equal to the number of prediction methods in \code{x}.
}
\description{
Bootstrap observation resampling for triptych objects
}
\examples{
# Triptych construction
predictions <- matrix(runif(300), ncol = 3)
colnames(predictions) <- c("Method_1", "Method_2", "Method_3")
observations <- rbinom(100, 1, predictions[, 1])
tr <- triptych(predictions, observations)

resampling_cases(tr$murphy, n_boot = 50)
resampling_cases(tr$reliability, position = "estimate", n_boot = 50)
resampling_cases(tr$roc, n_boot = 50)

}
